See the message
So at the !!!!Here!!! you should explain about which source code repo it is using and what is in the source code repo. I don't know FUSE and so as a new developer I also need somewhat of an introduction to FIS. I know it will make this longer but I think it is critical if you want people to understand FIS.

Also, it would be good to give a couple of sentences on what kind of application we are building. It could be in the chapter before. 

ALSO COMMENTS IN ALL CAPS